Article indexes as u64 instead of u32
The RFC specifies that article counts fit in a signed int, but the limit
has been surpassed in some newsgroups, and servers are returning numbers
that would fit in a u64.

+++ src/types/mod.rs
@@ -20,8 +20,8 @@ pub mod response_code;
 /// The number of an article relative to a specific Newsgroup
 ///
 /// Per [RFC 3977](https://tools.ietf.org/html/rfc3977#section-6) article numbers should fit within
-/// 31-bits.
-pub type ArticleNumber = u32;
+/// 31-bits but has been surpassed since.
+pub type ArticleNumber = u64;
 
 /// Re-exports of traits and response types
 pub mod prelude {
